build-mariner-arm64: build ISO using custom kernel
This commit uses Chris' fork of the CBL-MarinerDemo repo which has a new
spec file for a "custom-kernel" package. This custom-kernel package is a
copy of the standard Mariner 2.0 kernel, with a couple minor adjustments
due to the new name.
The repo also updates the Demo ISO image configuration to use the
custom-kernel package instead of our standard kernel package.

Signed-off-by: Chris Co <chrco@microsoft.com>